backend/requirements.txt
flask
openai
python-dotenv==1.0.0
